Changed the structure of the files to a more modular format. By removing utils and creating a data folder, the code is now more organized and easier to navigate.
